moved Base Logger functionality into its own file
